pyproject.toml
[tool.poetry]
name = "grid"
version = "0.1.0"
description = "Grid tracker"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
gunicorn = "^20.1.0"
celery = {extras = ["sqs"], version = "^5.3.1"}
djoser = "^2.2.0"
django-cors-headers = "^4.2.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
